// src/presentation/http/docs/resolved-docs-options.interface.ts
import type { OpenApiServer } from "./openapi-server.interface.ts";
import type { ScalarOptions } from "./scalar-options.interface.ts";

export interface ResolvedDocsOptions {
  readonly enabled: boolean;
  readonly title: string;
  readonly version: string;
  readonly description?: string;
  readonly path: string;
  readonly openApiPath: string;
  readonly servers: OpenApiServer[];
  readonly scalar: Required<ScalarOptions>;
}